requirements.txt
torch		#==2.8.0
torchvision	#==0.23.0
scikit-image	#==0.25.2
pyyaml		#==6.0.2
scipy		#==1.15.3
pillow		#==11.3.0
